Modularize permutation circuit within crate
In order to implement the padding and the full keccak circuit, the
circuit for the permutation as well as the gates have been moved to it's
own module `permutation` so that everything has a distinct
responsability and place.
