--- keccak256/src/gates.rs
@@ -1,17 +1,0 @@
-#![allow(dead_code)]
-#![allow(clippy::type_complexity)]
-#![allow(clippy::too_many_arguments)]
-pub mod absorb;
-pub mod base_conversion;
-pub mod gate_helpers;
-pub mod iota_b13;
-pub mod iota_b9;
-pub mod mixing;
-pub mod pi;
-pub mod rho;
-pub mod rho_checks;
-pub mod rho_helpers;
-pub mod state_conversion;
-pub mod tables;
-pub mod theta;
-pub mod xi;
