+++ keccak256/src/lib.rs
@@ -2,9 +2,8 @@
 // just used in tests
 
 pub mod arith_helpers;
-pub mod circuit;
 pub mod common;
-pub mod gates;
+pub mod permutation;
 // We build arith module to get test cases for the circuit
 pub mod keccak_arith;
 // We build plain module for the purpose of reviewing the circuit
