+++ keccak256/src/permutation.rs
@@ -1,0 +1,18 @@
+#![allow(dead_code)]
+#![allow(clippy::type_complexity)]
+#![allow(clippy::too_many_arguments)]
+pub(crate) mod absorb;
+pub(crate) mod base_conversion;
+pub mod circuit;
+pub(crate) mod gate_helpers;
+pub(crate) mod iota_b13;
+pub(crate) mod iota_b9;
+pub(crate) mod mixing;
+pub(crate) mod pi;
+pub(crate) mod rho;
+pub(crate) mod rho_checks;
+pub(crate) mod rho_helpers;
+pub(crate) mod state_conversion;
+pub(crate) mod tables;
+pub(crate) mod theta;
+pub(crate) mod xi;
