+++ keccak256/src/permutation/absorb.rs
