+++ keccak256/src/permutation/base_conversion.rs
@@ -4,7 +4,7 @@ use halo2_proofs::{
     poly::Rotation,
 };
 
-use crate::gates::tables::BaseInfo;
+use super::tables::BaseInfo;
 use eth_types::Field;
 
 #[derive(Clone, Debug)]
@@ -163,7 +163,7 @@ impl<F: Field> BaseConversionConfig<F> {
 mod tests {
     use super::*;
     use crate::arith_helpers::{convert_b2_to_b13, convert_b9_lane_to_b13};
-    use crate::gates::{
+    use crate::permutation::{
         gate_helpers::biguint_to_f,
         tables::{FromBase9TableConfig, FromBinaryTableConfig},
     };
