+++ keccak256/src/permutation/gate_helpers.rs
