+++ keccak256/src/permutation/iota_b13.rs
@@ -197,7 +197,7 @@ impl<F: Field> IotaB13Config<F> {
 mod tests {
     use super::*;
     use crate::common::{PERMUTATION, ROUND_CONSTANTS};
-    use crate::gates::gate_helpers::biguint_to_f;
+    use crate::permutation::gate_helpers::biguint_to_f;
     use halo2_proofs::circuit::Layouter;
     use halo2_proofs::plonk::{Advice, Column, ConstraintSystem, Error};
     use halo2_proofs::{circuit::SimpleFloorPlanner, dev::MockProver, plonk::Circuit};
