+++ keccak256/src/permutation/rho_helpers.rs
